+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 8.21
 -----
-
+*   Bug Fixes
+    *   Keep podcast artwork available offline by caching it more reliably
+        ([#5549](https://github.com/Automattic/pocket-casts-android/pull/5549))
 
 8.20
 -----